improve it, but it still wrong

+++ cycled_priority_queue.c
@@ -5,12 +5,25 @@
 int init(cpq_t *q) {
     q->head = NULL;
     return 0;
-};
+}
+
+cpq_node_t *initNode(const cpq_t *q, void *data) {
+    cpq_node_t *node = (cpq_node_t *) malloc(sizeof(cpq_node_t));
+    node->data = data;
+    node->next = q->head;
+    node->prev = q->head;
+    return node;
+}
 
 cpq_node_ptr_t *selectEnqueueNode(const cpq_t *q, void **data) {
     cpq_node_ptr_t *current = q->head;
-    while (current->first->next != q->head && current->second->next != q->head) {
-        if ((int) current->first->data < (int) data) {
+    while (current->first->next != q->head &&
+            (current->second != NULL && current->second->next != q->head)) {
+        if ((int) current->first->data > (int) data) {
+            if (current->second == NULL) {
+                current->second = initNode(q, data);
+                return current;
+            }
             current = current->second->next;
         } else {
             current = current->first->next;
@@ -21,22 +34,15 @@ cpq_node_ptr_t *selectEnqueueNode(const cpq_t *q, void **data) {
 
 cpq_node_ptr_t *selectDequeueNode(const cpq_t *q) {
     cpq_node_ptr_t *current = q->head;
-    while (current->first->next != q->head && current->second->next != q->head) {
-        if ((int) current->first->data < (int) current->second->data) {
+    while (current->first->next != q->head &&
+           current->second != NULL && current->second->next != q->head) {
+        if ((int) current->first->data > (int) current->second->data) {
             current = current->second->next;
         } else {
             current = current->first->next;
         }
     }
     return current;
-}
-
-cpq_node_t *initNode(const cpq_t *q, void *data) {
-    cpq_node_t *node = (cpq_node_t *) malloc(sizeof(cpq_node_t));
-    node->data = data;
-    node->next = q->head;
-    node->prev = q->head;
-    return node;
 }
 
 void insertFirst(cpq_node_ptr_t *node, cpq_node_ptr_t *nodeToInsert) {
@@ -58,29 +64,32 @@ int enqueue(cpq_t *q, void *data) {
 
     node->second = (cpq_node_t *) malloc(sizeof(cpq_node_t));
     node->first = initNode(q, data);
-    node->second = initNode(q, data);
+    node->second = NULL;
+//    node->second = initNode(q, data);
 
     if (q->head == NULL) {
         q->head = node;
         q->head->first->next = q->head;
         q->head->first->prev = q->head;
 
-        q->head->second->next = q->head;
-        q->head->second->prev = q->head;
+//        q->head->second->next = q->head;
+//        q->head->second->prev = q->head;
         return 0;
     }
 
     cpq_node_ptr_t *nodeToInsert = selectEnqueueNode(q, data);
 
     insertFirst(node, nodeToInsert);
-    insertSecond(node, nodeToInsert);
+    if (node->second != NULL) {
+        insertSecond(node, nodeToInsert);
+    }
 
     return 0;
 };
 
 void deleteLinks(const cpq_node_t *nodeToDelete) {
-    nodeToDelete->prev->first->next = nodeToDelete->next;
-    nodeToDelete->next->first->prev = nodeToDelete->prev;
+//    nodeToDelete->prev->first->next = nodeToDelete->next;
+//    nodeToDelete->next->first->prev = nodeToDelete->prev;
 }
 
 int dequeue(cpq_t *q, void **data) {
@@ -90,26 +99,34 @@ int dequeue(cpq_t *q, void **data) {
 
     deleteLinks(nodeToDelete->first);
 
-    deleteLinks(nodeToDelete->second);
+    if (nodeToDelete->second != NULL) {
+        deleteLinks(nodeToDelete->second);
+    }
 
     *data = nodeToDelete->first->data;
-    *data = nodeToDelete->second->data;
+    if (nodeToDelete->second != NULL) {
+        *data = nodeToDelete->second->data;
+    }
+
     free(nodeToDelete);
 
     return 0;
-};
+}
 
 void printList(cpq_t q) {
     cpq_node_ptr_t *current = q.head;
+    int i = 0;
     do {
         printf(" %d ", (int) current->first->data);
-        printf(" %d ", (int) current->second->data);
+        if (current->second != NULL) {
+            printf(" %d ", (int) current->second->data);
+        }
 
-        if ((int) current->first->data < (int) current->second->data) {
+        if (current->second != NULL && ((int) current->first->data < (int) current->second->data)) {
             current = current->second->next;
         } else {
             current = current->first->next;
         }
-        current = current->first->next;
-    } while (current != q.head);
+        ++i;
+    } while (current != q.head && i<10);
 }
